requirements.txt
python==3.9
pymongo==4.8.0
scikit-learn==1.5.1
tensorflow==2.17.0
nltk==3.8.1
pandas==2.2.2
numpy==1.26.4
flask==3.0.3
python-dotenv==1.0.1
faker==20.1.0